refactor: move test-lib.js to test folder
So it's handled as dev context
