+++ jest.config.js
@@ -6,6 +6,6 @@ module.exports = {
   verbose: true,
   bail: false,
   testEnvironment: 'node',
-  testMatch: ['**/test/**/*.js'],
+  testMatch: ['**/test/**/*.js', '!**/test/test-lib.js'],
   setupTestFrameworkScriptFile: path.resolve('./jest.setupEnvironment.js'),
 };
