+++ test/test-lib.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const { expect } = require('chai');
-const main = require('./index');
-const ValidationError = require('./lib/ValidationError');
-const { transform } = require('./utils/transform-unicode-patterns');
+const main = require('../index');
+const ValidationError = require('../lib/ValidationError');
+const { transform } = require('../utils/transform-unicode-patterns');
 
 function generateTestCases(schema, positive, cases) {
   if (positive) {
